+++ src/main/java/emu/grasscutter/game/managers/energy/EnergyDropEntry.java
@@ -1,4 +1,4 @@
-package emu.grasscutter.game.managers.EnergyManager;
+package emu.grasscutter.game.managers.energy;
 
 import java.util.List;
