+++ src/main/java/emu/grasscutter/game/managers/energy/SkillParticleGenerationInfo.java
@@ -1,4 +1,4 @@
-package emu.grasscutter.game.managers.EnergyManager;
+package emu.grasscutter.game.managers.energy;
 
 public class SkillParticleGenerationInfo {
 	private int value;
