+++ src/main/java/emu/grasscutter/server/game/GameServer.java
@@ -10,10 +10,10 @@ import emu.grasscutter.game.drop.DropManager;
 import emu.grasscutter.game.dungeons.DungeonManager;
 import emu.grasscutter.game.expedition.ExpeditionManager;
 import emu.grasscutter.game.gacha.GachaManager;
-import emu.grasscutter.game.managers.ChatManager.ChatManager;
-import emu.grasscutter.game.managers.ChatManager.ChatManagerHandler;
 import emu.grasscutter.game.managers.InventoryManager;
 import emu.grasscutter.game.managers.MultiplayerManager;
+import emu.grasscutter.game.managers.chat.ChatManager;
+import emu.grasscutter.game.managers.chat.ChatManagerHandler;
 import emu.grasscutter.game.player.Player;
 import emu.grasscutter.game.quest.ServerQuestHandler;
 import emu.grasscutter.game.shop.ShopManager;
