+++ src/main/java/emu/grasscutter/server/packet/send/PacketMarkMapRsp.java
@@ -1,6 +1,6 @@
 package emu.grasscutter.server.packet.send;
 
-import emu.grasscutter.game.managers.MapMarkManager.MapMark;
+import emu.grasscutter.game.managers.mapmark.MapMark;
 import emu.grasscutter.game.player.Player;
 import emu.grasscutter.net.packet.BasePacket;
 import emu.grasscutter.net.packet.PacketOpcodes;
